update totalAmount in adding order

+++ Book-service/src/main/java/com/samah/bookservice/dto/BookDto.java
@@ -19,7 +19,7 @@ public class BookDto {
     private List<String> authors;
     private String isbn;
     private List<String> genres;
-    private BigDecimal price;
+    private double price;
     private int quantity;
     private int publicationYear;
     private String publisher;
